backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=24.0.0
